implement asyc sha256 hasher, create new utils.py
